+++ src/level/objects/shelf.cpp
@@ -53,7 +53,7 @@ InstanceDynamicObject::GoToData ObjectShelf::GetGoToData(InstanceDynamicObject* 
 
       if (waypoint1 && (_level->FindPath(path, *waypoint, *waypoint1)))
       {
-	if (ret.max_distance > path.size() || ret.max_distance == -1)
+	if (ret.max_distance > (int)path.size() || ret.max_distance == -1)
 	{
 	  ret.nearest      = waypoint1;
 	  ret.max_distance = path.size();
@@ -62,7 +62,7 @@ InstanceDynamicObject::GoToData ObjectShelf::GetGoToData(InstanceDynamicObject* 
 
       if (waypoint2 && (_level->FindPath(path, *waypoint, *waypoint2)))
       {
-	if (ret.max_distance > path.size() || ret.max_distance == -1)
+	if (ret.max_distance > (int)path.size() || ret.max_distance == -1)
 	{
 	  ret.nearest      = waypoint2;
 	  ret.max_distance = path.size();
